# wgengine/filter: split "local IP" lookups by IPv4-vs-IPv6