@bradfitz 2024-06-16
I'll wait for https://github.com/tailscale/tailscale/pull/12493 to be in first and then fix the commit message benchmarks here with the better names.