@bradfitz 2024-06-16
Closing this for now. After all the other changes landed, I'm not seeing any benchmark wins from this now.