@bradfitz 2024-06-16
Actually this might help with #12497, so reopening. Need to run more benchmarks.